requirements.txt
urlextract
requests
lxml
prometheus_client
tornado
jinja2
peewee
pytube
mutagen
user_agent
pytelegrambotapi
unidecode
python-vlc
discord.py
discord.py[voice]
python-dotenv
